prepbytes searching questions

+++ general/sort/MergeSort.java
@@ -2,7 +2,7 @@ package general.sort;
 
 public class MergeSort {
 
-	private static void mergeSort(long[] arr, int left, int right) {
+	private static void mergeSort(int[] arr, int left, int right) {
 		if (left < right) {
 			int middle = (left + right) / 2;
 			mergeSort(arr, left, middle);
@@ -11,15 +11,15 @@ public class MergeSort {
 		}
 	}
 
-	private static void merge(long[] arr, int left, int middle, int right) {
-		long[] t1 = new long[middle - left + 2];
-		long[] t2 = new long[right - middle + 1];
+	private static void merge(int[] arr, int left, int middle, int right) {
+		int[] t1 = new int[middle - left + 2];
+		int[] t2 = new int[right - middle + 1];
 		for (int i = 0; i <= middle - left; i++)
 			t1[i] = arr[left + i];
 		for (int i = 0; i < right - middle; i++)
 			t2[i] = arr[middle + i + 1];
-		t1[t1.length - 1] = Long.MAX_VALUE;
-		t2[t2.length - 1] = Long.MAX_VALUE;
+		t1[t1.length - 1] = Integer.MAX_VALUE;
+		t2[t2.length - 1] = Integer.MAX_VALUE;
 		int i1 = 0, i2 = 0;
 		for (int i = left; i <= right; i++) {
 			if (t1[i1] < t2[i2])
